feat: changed location for common utilites
