+++ src/DotNetNuke.PowerBI/Views/ContentView/Index.cshtml
@@ -315,7 +315,7 @@
                                 <td><strong>@Dnn.LocalizeString("SubscriptionEmailSubject")</strong> <span data-bind="text: emailSubject"></span></td>
                             </tr>
                             <tr>
-                                <td><strong>@Dnn.LocalizeString("SubscriptionMessage")</strong> <span data-bind="text: message"></span></td>
+                                <td><strong>@Dnn.LocalizeString("SubscriptionEmailMessage")</strong> <span data-bind="text: message"></span></td>
                             </tr>
                             <tr>
                                 <td><strong>@Dnn.LocalizeString("SubscriptionReportPages")</strong></td>
